+++ tests/next-16/app/page.tsx
@@ -1,0 +1,23 @@
+import { Suspense } from 'react';
+import { cookieFlag, exampleFlag, hostFlag } from '../flags';
+
+async function Home() {
+  const example = await exampleFlag();
+  const host = await hostFlag();
+  const cookie = await cookieFlag();
+  return (
+    <div>
+      <h1>Example App Router Flag Value: {example ? 'true' : 'false'}</h1>
+      <p>Host: {host}</p>
+      <p>Cookie: {cookie}</p>
+    </div>
+  );
+}
+
+export default function Page() {
+  return (
+    <Suspense>
+      <Home />
+    </Suspense>
+  );
+}