+++ packages/adapter-growthbook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flags-sdk/growthbook
 
+## 0.1.1
+
+### Patch Changes
+
+- ff052f0: upgrade internal @vercel/edge-config dependency to v1.4.3
+
 ## 0.1.0
 
 ### Minor Changes